Implement `q_sort()`, but need to optimize
also revise allocation in `q_insert_head`, `q_insert_tail`, wont't try
to allocate memory if `q` is NULL

+++ queue.h
@@ -26,10 +26,7 @@ typedef struct ELE {
 typedef struct {
     list_ele_t *head; /* Linked list of elements */
     list_ele_t *tail;
-    /* TODO: You will need to add more fields to this structure
-     *        to efficiently implement q_size and q_insert_tail.
-     */
-    int size;
+    size_t size;
 } queue_t;
 
 /* Operations on queue */
